update yyyyMmDdDate type
