+++ app/views/portal/index.html.erb
@@ -79,7 +79,7 @@
         description: null,
         options: {},
         is_hero: true,
-        last_updated_at: '2016-09-06 05:28:50.365576',
+        last_updated_at: "2015-02-03T00:00:00.000Z",
         published_at: '2016-09-06 05:28:50.356717',
         created_at: '2016-09-06 05:28:50.356717',
         updated_at: '2016-09-06 05:28:50.366554',
@@ -97,7 +97,7 @@
         description: 'Overall satisfaction score includes all ratings weighted from 100% for very satisfied to 0% for very dissatisfied',
         options: {},
         is_hero: false,
-        last_updated_at: '2016-09-06 05:28:50.365576',
+        last_updated_at: "2015-02-03T00:00:00.000Z",
         published_at: '2016-09-06 05:28:50.356717',
         created_at: '2016-09-06 05:28:50.356717',
         updated_at: '2016-09-06 05:28:50.366554'
